backend/requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Image Processing
Pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0

# Utilities
pydantic>=2.0.0
aiofiles>=23.0.0

# CORS
starlette>=0.27.0

